05_web_security/15_docker_swarm/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Werkzeug
